Moving src paths
