+++ workspace/src/github.com/previousnext/k8s-ssh-server/cli/main.go
@@ -7,7 +7,7 @@ import (
 	"github.com/gosuri/uitable"
 	"k8s.io/client-go/rest"
 
-	sshclient "github.com/previousnext/client"
+	sshclient "github.com/previousnext/k8s-ssh-server/client"
 )
 
 func main() {
