+++ workspace/src/github.com/previousnext/k8s-ssh-server/client/types.go
